requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
jinja2==3.1.3
googletrans==4.0.0-rc1
deep-translator==1.11.4
openpyxl==3.1.2
